Improved logging, error handling, etc.

+++ Torrent.php
@@ -8,58 +8,38 @@ class Torrent extends Module
 
   public function derive()
   {
-    // the following class variables (among others - see Module.inc) will already be set
-    // when this function gets called:
-    // $this->identifier   the item ID
-    // $this->itemDir      full path to item dir, with trailing /
-
-    // $this->sourceFile   full path to source file (*.torrent, in this case); may be in subdir of itemdir
-    // $this->sourceFormat source format name (as given in <dependency ... variation="{name}" /> in derivations.xml)
-
-    // $this->targetFile   likewise for the target file
-    // $this->targetFormat (as given in <variation name="{name}" ... /> in derivations.xml)
-
-    // $this->tmp;         preexisting empty tmp dir (with trailing /) for any files you need
-    //                     to create (will later be rm'd automatically)
-
     // REQUIRES:
-    //  transmission-daemon   configuration: listening on default port 9091 for rpc from transmission-remote CLI
+    //  transmission-daemon   current configuration: listening on default port 9091 for rpc from transmission-remote CLI
     //  transmission-remote   
     
+    // INPUTS:
+    //   foo.torrent
     // OUTPUTS:
-    //   .torrentcontents   comma delimited manifest of files in .torrent:  filename,size in bytes 
-    //   torrent contents 
-    
-    // TODO:
-    //   check for non-conforming filenames (e.g. illegal characters leading to exploits)
-    
+    //   foo_torrent.tlg        comma delimited manifest of files in .torrent:  orig filename,sanitzed filename,size in bytes 
+    //   <torrent contents>     note that filenames are changed to Archive-safe versions (no UTF support currently!)
+        
     // NOTES:
+    //   currently uses one daemon, we will move to one daemon/session
     //   currently assumes one torrent per item, as with scanned books (but what about .torrents full of .torrents?)
-    //   currently using transmission-daemon's whitelist for IP ranges from which it accepts control
-    //   currently moving .torrent.log file containing output of btget, could discard, or is it of historic interest?
-    
-    // transmission-daemon needs write permission, it runs under debian-transmission by default :P
-    Util::cmd('chmod ugo+w '.Util::esc($this->tmp), 'PRINT');    
-    
+
     // DEBUG
-    Util::cmdPP('echo | whereis transmission-remote'); // last check still not installed on workers, so running out of /home/ximmm :P
+    // Util::cmdPP('echo | whereis transmission-remote');
     // Util::cmdPP('echo | whereis transmission-daemon');
-    // push out the revised permissions in /home/ximm/petabox/etc/transmission-daemon/settings.json 
-    // Util::cmd('sudo cp /etc/transmission-daemon/settings.json /home/ximm/projects/bitty/bitlogs/default_settings.json', 'PRINT');
     
-    // This is now hopefully unnecessary since Raj is pushing into /petabox/etc/transmission-daemon/settings.json to all workers
-    // Util::cmdPP('sudo cp /home/ximm/projects/bitty/bitlogs/settings.json /etc/transmission-daemon/settings.json');
-    Util::cmdPP('sudo /etc/init.d/transmission-daemon reload');
-
-    // TODO: move script into petabox bin basket 
-    Util::cmd('/home/ximm/projects/bitty/btget -stdout -log=/home/ximm/projects/bitty/bitlogs/btgettest.log '.Util::esc($this->sourceFile).' -dir='.Util::esc($this->tmp), 'PRINT');   
-
-    // seize ownership of downloaded files; may not be necessary in worker environment, not sure what user 
-    //  derive.php is executed under
-    // Util::cmd('sudo chown -R ximm '.Util::esc($this->tmp), 'PRINT');        
-    Util::cmd('sudo chmod ugo+w '.Util::esc($this->tmp).'*', 'PRINT');
+    $btget = configGetPetaboxPath('bin-btget');
     
-    // in addition to the manifest, I generate a disposable helper file including:
+    // give transmission-daemon write permission, it runs under debian-transmission by default
+    // when we are starting/stopping daemon for each session this may no longer be necessary
+    Util::cmd('chmod ugo+w '.Util::esc($this->tmp), 'PRINT');    
+     
+    // Make sure settings are current on the local daemon. settings.json symlinked from /petabox/etc/transmisison-daemon
+    // TODO: start daemon here with a specific RPC port and peer port
+    Util::cmdPP('/etc/init.d/transmission-daemon reload');
+
+    // Util::cmdPP(Util::esc($_SERVER['PETABOX_HOME']).'/sw/bin/btget.py -stdout -sanitize -verbose '.Util::esc($this->sourceFile).' -dir='.Util::esc($this->tmp) );
+    Util::cmdPP("$btget -stdout -sanitize -verbose ".Util::esc($this->sourceFile)." -dir=".Util::esc($this->tmp) );   
+    
+    // in addition to the manifest, btget generates a disposable helper file the contents of which are:
     //  (a) the SHA1 infohash fingerprint for the torrent; and 
     //  (b) the torrent name field from its bencoded info, which == the directory name created by transmission-daemon
     $targetHashFile = $this->tmp.$this->shortName($this->sourceFile).'hash';        
@@ -68,43 +48,62 @@ class Torrent extends Module
         $torrentHash = strtok($lines[0], "\n");
         $torrentName = strtok($lines[1], "\n");
     } else {
-        Util::cmdPP('echo MISSING HASH FILE:'. Util::esc($targetHashFile) );
+        echo "MISSING HASH FILE: $targetHashFile\n";
     }
     
-    // torrent contents are in $this->tmp/$torrentName/... while moving eliminate the intermediate dir $torrentName
-    $contentPath = $this->tmp.$torrentName;    
-    Util::cmd('cp -r '.Util::esc($contentPath).'/* '.Util::esc($this->itemDir), 'PRINT');    
-    Util::cmd('rm -rf '.Util::esc($contentPath), 'PRINT');    
+    // torrent contents are automatically stored by transmission-d in $this->tmp/$torrentName/...
+    // this path is a subdirectory for multi-file torrents, but a single file only for single-file torrents
+    // while moving, eliminate the intermediate dir $torrentName
+    // NOTE: $torrentName is not a sanized filename, but we fail on retrieval if it appears malicious
+    $contentPath = $this->tmp.$torrentName;
+    if (is_dir($contentPath) ) {
+        Util::cmd('cp -r '.Util::esc($contentPath).'/* '.Util::esc($this->itemDir), 'PRINT');    
+        Util::cmd('rm -rf '.Util::esc($contentPath), 'PRINT');
+    } else {
+        // single-file case 
+        Util::cmd('mv '.Util::esc($contentPath).' '.Util::esc($this->itemDir), 'PRINT');    
+    }    
     
     // move the generated manifest file which is our targetFile
     $targetContentsFile = $this->tmp.$this->shortName($this->targetFile);
     Util::cmd('mv '.Util::esc($targetContentsFile).' '.Util::esc($this->itemDir), 'PRINT');    
 
+    // move the log file if there is one
+    $targetLogFile = $this->tmp.$this->shortName($this->sourceFile).'.log';        
+//    if (file_exists($targetLogFile)) {
+//        Util::cmd('mv '.Util::esc($targetLogFile).' '.Util::esc($this->itemDir), 'PRINT');    
+//    }
+
     // update _meta.xml 
-    ModifyXML::updateElem('source', $this->shortName($this->sourceFile),
+    // TODO: use ModifyXML per Hank's guidance to support multiple .torrents per item; get Alexis' sign off on tags
+//    ModifyXML::updateElem('source', $this->shortName($this->sourceFile),
+//                          "{$this->identifier}_meta.xml", $this->tmp);        
+    ModifyXML::updateElem('external-identifier', "torrent:urn:sha1:".$torrentHash,
                           "{$this->identifier}_meta.xml", $this->tmp);        
-    ModifyXML::updateElem('identifier-torrent-infohash', $torrentHash,
-                          "{$this->identifier}_meta.xml", $this->tmp);        
-    ModifyXML::updateElem('identifier-torrent-name', $torrentName,
-                          "{$this->identifier}_meta.xml", $this->tmp);        
+//    ModifyXML::updateElem('identifier-torrent-name', "torrent:filename:".$torrentName,
+//                          "{$this->identifier}_meta.xml", $this->tmp);        
                           
+    // target manifest file is line-delimited information for each file in the torrent, in the format
+    //  originalname,sanitizedname,length
+    
     if (file_exists($this->targetFile)) {
         $fg = new FormatGetter;
         $lines = file($this->targetFile);
         foreach ($lines as $line) {
-            $outFile = strtok($line, ',');
+            $originalFile = strtok($line, ',');
+            $outFile = strtok(',');
             $outAbsPath = $this->itemDir.$outFile;
             if (file_exists($outAbsPath)) {
                 $formatName = $fg->pickFormat($outFile);
                 $this->extraTarget($outFile, $formatName);
             } else {
                 // file in .torrent missing from download directory
-                Util::cmdPP('echo MISSING FILE FROM TORRENT:' . Util::esc($outAbsPath) );
+                echo "MISSING FILE FROM TORRENT: $outAbsPath\n";
             }
         }
     } else {  
         // manifest file missing
-        Util::cmdPP('echo MISSING TARGET FILE: '.Util::esc($manifest) );
+        echo "MISSING TARGET FILE: $manifest\n";
         fatal('Target manifest for torrent missing: '.Util::esc($manifest));
     }
